backend/python/requirements.txt
fastapi
uvicorn
transformers
torch
numpy
emoji
soynlp
scikit-learn
python-dotenv
dicttoxml
gunicorn


# MkDocs
mkdocs
mkdocs-get-deps
mkdocs-i18n
mkdocs-material
mkdocs-material-extensions
mkdocs-rtd-dropdown
mkdocs-static-i18n
mkdocs-toc-md